+++ examples/quickstarts/groups_members/resource_collection/main.tf
@@ -63,7 +63,7 @@ resource "msgraph_resource" "user3" {
 # Manage all group members as a collection (RECOMMENDED APPROACH)
 # This provides declarative management where you specify the complete desired state
 resource "msgraph_resource_collection" "group_members" {
-  url           = "groups/${msgraph_resource.group.id}/members/$ref"
+  url = "groups/${msgraph_resource.group.id}/members/$ref"
   reference_ids = [
     msgraph_resource.user1.id,
     msgraph_resource.user2.id,