+++ CHANGELOG.md
@@ -10,6 +10,9 @@ DEPENDENCIES:
 - Updated `github.com/Azure/azure-sdk-for-go/sdk/azidentity` from v1.8.0 to v1.13.0 to enable Azure PowerShell authentication support
 - Updated `github.com/Azure/azure-sdk-for-go/sdk/azcore` from v1.16.0 to v1.19.1
 
+BUG FIXES:
+- Fixed an issue where `@odata.type` property was missing in PATCH requests for resources that require it (e.g. Named Locations) ([#59](https://github.com/microsoft/terraform-provider-msgraph/issues/59))
+
 ## 0.2.0
 
 FEATURES: